Module aws_sdk_panorama::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

CreateApplicationInstance

Fluent builder constructing a request to CreateApplicationInstance.

CreateJobForDevices

Fluent builder constructing a request to CreateJobForDevices.

CreateNodeFromTemplateJob

Fluent builder constructing a request to CreateNodeFromTemplateJob.

CreatePackage

Fluent builder constructing a request to CreatePackage.

CreatePackageImportJob

Fluent builder constructing a request to CreatePackageImportJob.

DeleteDevice

Fluent builder constructing a request to DeleteDevice.

DeletePackage

Fluent builder constructing a request to DeletePackage.

DeregisterPackageVersion

Fluent builder constructing a request to DeregisterPackageVersion.

DescribeApplicationInstance

Fluent builder constructing a request to DescribeApplicationInstance.

DescribeApplicationInstanceDetails

Fluent builder constructing a request to DescribeApplicationInstanceDetails.

DescribeDevice

Fluent builder constructing a request to DescribeDevice.

DescribeDeviceJob

Fluent builder constructing a request to DescribeDeviceJob.

DescribeNode

Fluent builder constructing a request to DescribeNode.

DescribeNodeFromTemplateJob

Fluent builder constructing a request to DescribeNodeFromTemplateJob.

DescribePackage

Fluent builder constructing a request to DescribePackage.

DescribePackageImportJob

Fluent builder constructing a request to DescribePackageImportJob.

DescribePackageVersion

Fluent builder constructing a request to DescribePackageVersion.

ListApplicationInstanceDependencies

Fluent builder constructing a request to ListApplicationInstanceDependencies.

ListApplicationInstanceNodeInstances

Fluent builder constructing a request to ListApplicationInstanceNodeInstances.

ListApplicationInstances

Fluent builder constructing a request to ListApplicationInstances.

ListDevices

Fluent builder constructing a request to ListDevices.

ListDevicesJobs

Fluent builder constructing a request to ListDevicesJobs.

ListNodeFromTemplateJobs

Fluent builder constructing a request to ListNodeFromTemplateJobs.

ListNodes

Fluent builder constructing a request to ListNodes.

ListPackageImportJobs

Fluent builder constructing a request to ListPackageImportJobs.

ListPackages

Fluent builder constructing a request to ListPackages.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

ProvisionDevice

Fluent builder constructing a request to ProvisionDevice.

RegisterPackageVersion

Fluent builder constructing a request to RegisterPackageVersion.

RemoveApplicationInstance

Fluent builder constructing a request to RemoveApplicationInstance.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateDeviceMetadata

Fluent builder constructing a request to UpdateDeviceMetadata.