Enum aws_sdk_panorama::Error

#[non_exhaustive]
pub enum Error {
    AccessDeniedException(AccessDeniedException),
    ConflictException(ConflictException),
    InternalServerException(InternalServerException),
    ResourceNotFoundException(ResourceNotFoundException),
    ServiceQuotaExceededException(ServiceQuotaExceededException),
    ValidationException(ValidationException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),
}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDeniedException(AccessDeniedException)

The requestor does not have permission to access the target action or resource.

Tuple Fields of AccessDeniedException

0: AccessDeniedException

ConflictException(ConflictException)

The target resource is in use.

Tuple Fields of ConflictException

0: ConflictException

InternalServerException(InternalServerException)

An internal error occurred.

Tuple Fields of InternalServerException

0: InternalServerException

ResourceNotFoundException(ResourceNotFoundException)

The target resource was not found.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

ServiceQuotaExceededException(ServiceQuotaExceededException)

The request would cause a limit to be exceeded.

Tuple Fields of ServiceQuotaExceededException

0: ServiceQuotaExceededException

ValidationException(ValidationException)

The request contains an invalid parameter value.

Tuple Fields of ValidationException

0: ValidationException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>