Module aws_sdk_panorama::error

Errors that can occur when calling the service.

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

validation_exception

See ValidationException

Structs

AccessDeniedException

The requestor does not have permission to access the target action or resource.

ConflictException

The target resource is in use.

CreateApplicationInstanceError

Error type for the CreateApplicationInstance operation.

CreateJobForDevicesError

Error type for the CreateJobForDevices operation.

CreateNodeFromTemplateJobError

Error type for the CreateNodeFromTemplateJob operation.

CreatePackageError

Error type for the CreatePackage operation.

CreatePackageImportJobError

Error type for the CreatePackageImportJob operation.

DeleteDeviceError

Error type for the DeleteDevice operation.

DeletePackageError

Error type for the DeletePackage operation.

DeregisterPackageVersionError

Error type for the DeregisterPackageVersion operation.

DescribeApplicationInstanceDetailsError

Error type for the DescribeApplicationInstanceDetails operation.

DescribeApplicationInstanceError

Error type for the DescribeApplicationInstance operation.

DescribeDeviceError

Error type for the DescribeDevice operation.

DescribeDeviceJobError

Error type for the DescribeDeviceJob operation.

DescribeNodeError

Error type for the DescribeNode operation.

DescribeNodeFromTemplateJobError

Error type for the DescribeNodeFromTemplateJob operation.

DescribePackageError

Error type for the DescribePackage operation.

DescribePackageImportJobError

Error type for the DescribePackageImportJob operation.

DescribePackageVersionError

Error type for the DescribePackageVersion operation.

InternalServerException

An internal error occurred.

ListApplicationInstanceDependenciesError

Error type for the ListApplicationInstanceDependencies operation.

ListApplicationInstanceNodeInstancesError

Error type for the ListApplicationInstanceNodeInstances operation.

ListApplicationInstancesError

Error type for the ListApplicationInstances operation.

ListDevicesError

Error type for the ListDevices operation.

ListDevicesJobsError

Error type for the ListDevicesJobs operation.

ListNodeFromTemplateJobsError

Error type for the ListNodeFromTemplateJobs operation.

ListNodesError

Error type for the ListNodes operation.

ListPackageImportJobsError

Error type for the ListPackageImportJobs operation.

ListPackagesError

Error type for the ListPackages operation.

ListTagsForResourceError

Error type for the ListTagsForResource operation.

ProvisionDeviceError

Error type for the ProvisionDevice operation.

RegisterPackageVersionError

Error type for the RegisterPackageVersion operation.

RemoveApplicationInstanceError

Error type for the RemoveApplicationInstance operation.

ResourceNotFoundException

The target resource was not found.

ServiceQuotaExceededException

The request would cause a limit to be exceeded.

TagResourceError

Error type for the TagResource operation.

UntagResourceError

Error type for the UntagResource operation.

UpdateDeviceMetadataError

Error type for the UpdateDeviceMetadata operation.

ValidationException

The request contains an invalid parameter value.

Enums

CreateApplicationInstanceErrorKind

Types of errors that can occur for the CreateApplicationInstance operation.

CreateJobForDevicesErrorKind

Types of errors that can occur for the CreateJobForDevices operation.

CreateNodeFromTemplateJobErrorKind

Types of errors that can occur for the CreateNodeFromTemplateJob operation.

CreatePackageErrorKind

Types of errors that can occur for the CreatePackage operation.

CreatePackageImportJobErrorKind

Types of errors that can occur for the CreatePackageImportJob operation.

DeleteDeviceErrorKind

Types of errors that can occur for the DeleteDevice operation.

DeletePackageErrorKind

Types of errors that can occur for the DeletePackage operation.

DeregisterPackageVersionErrorKind

Types of errors that can occur for the DeregisterPackageVersion operation.

DescribeApplicationInstanceDetailsErrorKind

Types of errors that can occur for the DescribeApplicationInstanceDetails operation.

DescribeApplicationInstanceErrorKind

Types of errors that can occur for the DescribeApplicationInstance operation.

DescribeDeviceErrorKind

Types of errors that can occur for the DescribeDevice operation.

DescribeDeviceJobErrorKind

Types of errors that can occur for the DescribeDeviceJob operation.

DescribeNodeErrorKind

Types of errors that can occur for the DescribeNode operation.

DescribeNodeFromTemplateJobErrorKind

Types of errors that can occur for the DescribeNodeFromTemplateJob operation.

DescribePackageErrorKind

Types of errors that can occur for the DescribePackage operation.

DescribePackageImportJobErrorKind

Types of errors that can occur for the DescribePackageImportJob operation.

DescribePackageVersionErrorKind

Types of errors that can occur for the DescribePackageVersion operation.

ListApplicationInstanceDependenciesErrorKind

Types of errors that can occur for the ListApplicationInstanceDependencies operation.

ListApplicationInstanceNodeInstancesErrorKind

Types of errors that can occur for the ListApplicationInstanceNodeInstances operation.

ListApplicationInstancesErrorKind

Types of errors that can occur for the ListApplicationInstances operation.

ListDevicesErrorKind

Types of errors that can occur for the ListDevices operation.

ListDevicesJobsErrorKind

Types of errors that can occur for the ListDevicesJobs operation.

ListNodeFromTemplateJobsErrorKind

Types of errors that can occur for the ListNodeFromTemplateJobs operation.

ListNodesErrorKind

Types of errors that can occur for the ListNodes operation.

ListPackageImportJobsErrorKind

Types of errors that can occur for the ListPackageImportJobs operation.

ListPackagesErrorKind

Types of errors that can occur for the ListPackages operation.

ListTagsForResourceErrorKind

Types of errors that can occur for the ListTagsForResource operation.

ProvisionDeviceErrorKind

Types of errors that can occur for the ProvisionDevice operation.

RegisterPackageVersionErrorKind

Types of errors that can occur for the RegisterPackageVersion operation.

RemoveApplicationInstanceErrorKind

Types of errors that can occur for the RemoveApplicationInstance operation.

TagResourceErrorKind

Types of errors that can occur for the TagResource operation.

UntagResourceErrorKind

Types of errors that can occur for the UntagResource operation.

UpdateDeviceMetadataErrorKind

Types of errors that can occur for the UpdateDeviceMetadata operation.