Module aws_sdk_panorama::model

Data structures used by operation inputs/outputs.

Modules

application_instance

See ApplicationInstance

conflict_exception_error_argument

See ConflictExceptionErrorArgument

device

See Device

device_job

See DeviceJob

device_job_config

See DeviceJobConfig

ethernet_payload

See EthernetPayload

ethernet_status

See EthernetStatus

job

See Job

job_resource_tags

See JobResourceTags

network_payload

See NetworkPayload

network_status

See NetworkStatus

node

See Node

node_from_template_job

See NodeFromTemplateJob

node_input_port

See NodeInputPort

node_instance

See NodeInstance

node_interface

See NodeInterface

node_output_port

See NodeOutputPort

ota_job_config

See OtaJobConfig

out_put_s3_location

See OutPutS3Location

package_import_job

See PackageImportJob

package_import_job_input_config

See PackageImportJobInputConfig

package_import_job_output

See PackageImportJobOutput

package_import_job_output_config

See PackageImportJobOutputConfig

package_list_item

See PackageListItem

package_object

See PackageObject

package_version_input_config

See PackageVersionInputConfig

package_version_output_config

See PackageVersionOutputConfig

s3_location

See S3Location

static_ip_connection_info

See StaticIpConnectionInfo

storage_location

See StorageLocation

validation_exception_error_argument

See ValidationExceptionErrorArgument

validation_exception_field

See ValidationExceptionField

Structs

ApplicationInstance

An application instance on a device.

ConflictExceptionErrorArgument

A conflict exception error argument.

Device

A device.

DeviceJob

A job that runs on a device.

DeviceJobConfig

A job's configuration.

EthernetPayload

A device's network configuration.

EthernetStatus

A device's Ethernet status.

Job

A job for a device.

JobResourceTags

Tags for a job.

NetworkPayload

The network configuration for a device.

NetworkStatus

The network status of a device.

Node

An application node that represents a camera stream, a model, code, or output.

NodeFromTemplateJob

A job to create a camera stream node.

NodeInputPort

A node input port.

NodeInstance

A node instance.

NodeInterface

A node interface.

NodeOutputPort

A node output port.

OtaJobConfig

An over-the-air update (OTA) job configuration.

OutPutS3Location

The location of an output object in Amazon S3.

PackageImportJob

A job to import a package version.

PackageImportJobInputConfig

A configuration for a package import job.

PackageImportJobOutput

Results of a package import job.

PackageImportJobOutputConfig

An output configuration for a package import job.

PackageListItem

A package summary.

PackageObject

A package object.

PackageVersionInputConfig

A package version input configuration.

PackageVersionOutputConfig

A package version output configuration.

S3Location

A location in Amazon S3.

StaticIpConnectionInfo

A static IP configuration.

StorageLocation

A storage location.

ValidationExceptionErrorArgument

A validation exception error argument.

ValidationExceptionField

A validation exception field.

Enums

ApplicationInstanceHealthStatus

ApplicationInstanceStatus

ConnectionType

DeviceConnectionStatus

DeviceStatus

DeviceType

JobResourceType

JobType

ManifestOverridesPayload

Parameter overrides for an application instance. This is a JSON document that has a

ManifestPayload

A application verion's manifest file. This is a JSON document that has a

NetworkConnectionStatus

NodeCategory

NodeFromTemplateJobStatus

NodeInstanceStatus

PackageImportJobStatus

PackageImportJobType

PackageVersionStatus

PortType

StatusFilter

TemplateType

UpdateProgress

ValidationExceptionReason